Move toNetwork methods to CommandUtils
