+++ src/main/java/cn/nukkit/command/CommandUtils.java
@@ -1,13 +1,20 @@
 package cn.nukkit.command;
 
 import cn.nukkit.Server;
+import cn.nukkit.command.data.CommandEnum;
+import cn.nukkit.command.data.CommandParamType;
+import cn.nukkit.command.data.CommandParameter;
 import cn.nukkit.locale.TextContainer;
 import cn.nukkit.locale.TranslationContainer;
 import cn.nukkit.permission.Permissible;
 import cn.nukkit.utils.TextFormat;
+import com.google.common.collect.ImmutableMap;
 import com.nukkitx.math.vector.Vector3f;
+import com.nukkitx.protocol.bedrock.data.CommandEnumData;
+import com.nukkitx.protocol.bedrock.data.CommandParamData;
 import lombok.experimental.UtilityClass;
 
+import java.util.Collections;
 import java.util.Optional;
 import java.util.Set;
 import java.util.regex.Matcher;
@@ -19,6 +26,25 @@ import static com.google.common.base.Preconditions.checkNotNull;
 @UtilityClass
 public class CommandUtils {
     private static final Pattern RELATIVE_PATTERN = Pattern.compile("(~)?(-?[0-9]*\\.?[0-9]*)");
+
+    private static final ImmutableMap<CommandParamType, CommandParamData.Type> NETWORK_TYPES = ImmutableMap.<CommandParamType, CommandParamData.Type>builder()
+            .put(CommandParamType.INT, CommandParamData.Type.INT)
+            .put(CommandParamType.FLOAT, CommandParamData.Type.FLOAT)
+            .put(CommandParamType.VALUE, CommandParamData.Type.VALUE)
+            .put(CommandParamType.WILDCARD_INT, CommandParamData.Type.WILDCARD_INT)
+            .put(CommandParamType.OPERATOR, CommandParamData.Type.OPERATOR)
+            .put(CommandParamType.TARGET, CommandParamData.Type.TARGET)
+            .put(CommandParamType.WILDCARD_TARGET, CommandParamData.Type.WILDCARD_TARGET)
+            .put(CommandParamType.FILE_PATH, CommandParamData.Type.FILE_PATH)
+            .put(CommandParamType.STRING, CommandParamData.Type.STRING)
+            .put(CommandParamType.POSITION, CommandParamData.Type.POSITION)
+            .put(CommandParamType.MESSAGE, CommandParamData.Type.MESSAGE)
+            .put(CommandParamType.TEXT, CommandParamData.Type.TEXT)
+            .put(CommandParamType.JSON, CommandParamData.Type.JSON)
+            .put(CommandParamType.COMMAND, CommandParamData.Type.COMMAND)
+            .put(CommandParamType.RAWTEXT, CommandParamData.Type.TEXT)
+
+            .build();
 
     public static Optional<Vector3f> parseVector3f(String[] args, Vector3f relative) {
         checkNotNull(args, "args");
@@ -47,6 +73,19 @@ public class CommandUtils {
             position += relative;
         }
         return position;
+    }
+
+    protected static CommandParamData toNetwork(CommandParameter commandParameter) {
+        return new CommandParamData(commandParameter.name, commandParameter.optional,
+                toNetwork(commandParameter.enumData), NETWORK_TYPES.get(commandParameter.type),
+                commandParameter.postFix, Collections.emptyList());
+    }
+
+    private static CommandEnumData toNetwork(CommandEnum commandEnum) {
+        if (commandEnum == null) {
+            return null;
+        }
+        return new CommandEnumData(commandEnum.getName(), commandEnum.getValues().toArray(new String[0]), false);
     }
 
     public static void broadcastCommandMessage(CommandSender source, String message) {
@@ -106,4 +145,5 @@ public class CommandUtils {
             }
         }
     }
+
 }
